for #2556, refactor ParserResultSetLoader
